Separate scss into individual files, create components folder
