+++ src/components/Header.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Header = () => {
+  return (
+    <div className="header-container">
+      <div className="header-title">Title here</div>
+    </div>
+  )
+}
+
+export default Header;
