+++ src/scss/_buttons.scss
@@ -1,0 +1,5 @@
+.btn-convert {
+  font-size: 1.75rem;
+  width: fit-content;
+  padding: 0.75rem 2rem;
+}
