+++ src/scss/_header.scss
@@ -1,0 +1,15 @@
+.header-container {
+  position: fixed;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  height: 4rem;
+  background-color: #eee;
+  width: 100%;
+  top: 0;
+  left: 0;
+}
+
+.header-title {
+  font-size: 2rem;
+}
