+++ src/scss/_input-recipe.scss
@@ -1,0 +1,11 @@
+.recipe-textarea {
+  display: block;
+  font-size: 1.2rem;
+  width: 100%;
+  max-width: 60rem;
+  min-height: 15rem;
+  resize: none;
+  border-radius: 0.5rem;
+  border: 2px solid rgb(91, 91, 91);
+  padding: 0.75rem;
+}
