+++ src/scss/_output-recipe.scss
@@ -1,0 +1,4 @@
+.converted-recipe-container {
+  width: 100%;
+  text-align: left;
+}
