+++ src/scss/main.scss
@@ -1,5 +1,7 @@
-body {
-  font-family: Arial, Helvetica, sans-serif;
-  background-color: #d9d9d9;
-  color: black;
-}
+@import "reset";
+
+// components
+@import "buttons";
+@import "header";
+@import "input-recipe";
+@import "output-recipe";
